+++ src/events.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::*;
+
+#[derive(Event, Default)]
+pub struct CollisionEvent;
