+++ src/main.rs
@@ -1,28 +1,21 @@
-#![allow(unused_imports)]
 use bevy::prelude::*;
 use bevy::window::{
     Window, WindowPlugin, WindowResolution,
 };
 use iyes_perf_ui::prelude::*;
-mod ball;
-mod config;
-mod game_state;
-mod net;
-mod player;
-mod score;
-mod setup;
-mod sound;
-mod wall;
 
-use ball::*;
-use config::{Config, get_window_mode, load_config};
+mod components;
+mod config;
+mod events;
+mod resources;
+mod game_state;
+mod prelude;
+mod systems;
+
+use prelude::*;
+
+use config::{get_window_mode, load_config};
 use game_state::*;
-use net::*;
-use player::*;
-use score::*;
-use setup::setup;
-use sound::*;
-use wall::*;
 
 fn main() {
     let config = load_config();
@@ -81,24 +74,25 @@ fn main() {
             ),
         ),
     );
-    app.add_event::<CollisionEvent>();
+    app.add_event::<events::CollisionEvent>();
 
     // systems
     app.add_systems(
-        Startup, setup,
+        Startup,
+        systems::setup,
     );
 
     app.add_systems(
         FixedUpdate,
         (
-            ball::apply_velocity,
-            ball::apply_gravity,
-            ball::apply_ball_bounds,
-            // ball::process_just_scored,
-            ball::check_for_ball_collisions,
-            sound::play_collision_sound,
-            player::player_movement,
-            player::apply_player_bounds,
+            systems::fixed_update::apply_velocity,
+            systems::fixed_update::apply_gravity,
+            systems::fixed_update::apply_ball_bounds,
+            // systems::fixed_update::process_just_scored,
+            systems::fixed_update::check_for_ball_collisions,
+            systems::fixed_update::play_collision_sound,
+            systems::fixed_update::player_movement,
+            systems::fixed_update::apply_player_bounds,
         )
             .chain()
             .in_set(GameplaySet),
@@ -110,7 +104,7 @@ fn main() {
     );
     app.add_systems(
         Update,
-        score::update_scoreboard
+        systems::update::update_scoreboard
             .run_if(in_state(GameModeState::Running)),
     );
     app.run();
