--- src/net.rs
@@ -1,4 +1,0 @@
-use bevy::prelude::*;
-
-#[derive(Component)]
-pub struct Net;
