+++ src/prelude.rs
@@ -1,0 +1,5 @@
+pub use bevy::prelude::*;
+pub use crate::components::*;
+pub use crate::events::*;
+pub use crate::resources::*;
+pub use crate::config::Config;
